fix(uplc): resolve clang-tidy findings in the UPLC sources
Clear the clang-tidy warnings the analysis build treats as errors: widen the find-first-set-bit arithmetic in the target type, merge identical switch and conditional branches into their default arms, drop top-level const from a forward-declared parameter, and align a forward declaration's parameter names with the definition.

+++ lib/src/uplc/builtins/builtins.c
@@ -2724,7 +2724,7 @@ body_find_first_set_bit(
         ++bit;
       }
 
-      result = (int64_t)((i * 8U) + bit);
+      result = ((int64_t)i * 8) + (int64_t)bit;
     }
   }
